Conflicts:
	AgeOfEmpiresTaller/VistaSrc/Vista.cpp

+++ AgeOfEmpiresTaller/VistaSrc/DibujoPersonaje.h
@@ -15,7 +15,7 @@
 #define SDL_RECT rect
 #define CANTIDAD_DE_MOVIMIENTOS 8
 
-enum movimientos {ABAJO,DIAGONAL_IZQUIERDA_ABAJO,IZQUIERDA,DIAGONAL_IZQUIERDA_ARRIBA,ARRIBA,DIAGONAL_DERECHA_ABAJO,DERECHA,DIAGONAL_DERECHA_ARRIBA};
+enum movimientos {ABAJO,DIAGONAL_IZQUIERDA_ABAJO,IZQUIERDA, DIAGONAL_IZQUIERDA_ARRIBA ,ARRIBA ,DIAGONAL_DERECHA_ABAJO,DERECHA, DIAGONAL_DERECHA_ARRIBA };
 
 using namespace std;
 class DibujoPersonaje: public Dibujo {
@@ -45,7 +45,7 @@ public:
 		this->movimiento_actual = mov;
 	}
 	void render(SDL_Renderer* renderer);
-	void elegir_frame(int des_x,int des_y);
+	void elegir_frame(double des_x,double des_y);
 	void cambiar_frame();
 	void setPixeles(int x, int y);
 	~DibujoPersonaje();
